UserLocationManager이름 수정, ISP 적용
load메서드 순서 변경
